requirements.txt
langchain
langchain-community
langchain-google-genai
chromadb
pypdf
streamlit
python-dotenv